requirements.txt
langchain-community==0.0.12
langchain-openai==0.0.2.post1
langchainhub==0.1.14
tabulate==0.9.0
streamlit-feedback==0.0.9
langchain-experimental==0.0.49
langchain==0.1.0
openai==1.6.1
openpyxl==3.1.2
pypdf==3.12.2
sentence-transformers==2.2.2
pydantic== 1.10.13
docarray== 0.40.0
tabulate==0.9.0
streamlit-feedback==0.0.9
numexpr==2.8.8
duckduckgo-search==4.0